Refactor emulator.rs into multiple submodules
This commit breaks down the large `src/term/emulator.rs` file into a
directory `src/term/emulator/` with several specialized modules,
improving organization and maintainability.

Key changes:
- The `TerminalEmulator` struct now resides in `src/term/emulator/mod.rs`.
- ANSI command processing logic moved to `ansi_handler.rs`.
- Backend and control event handling moved to `input_handler.rs`.
- Character printing and mapping moved to `char_processor.rs`.
- Screen manipulation operations (erase, scroll, etc.) moved to `screen_ops.rs`.
- SGR and terminal mode handling moved to `mode_handler.rs`.
- OSC command handling moved to `osc_handler.rs`.
- Cursor movement helpers and WindowManipulation logic moved to `cursor_handler.rs`.

The original `impl TerminalEmulator` block was split, with methods either
moved to these new modules or delegated from the original call sites.
The `src/term/emulator/methods.rs` file, which I created in an
earlier step, now holds remaining helper methods.

This refactoring aligns the codebase closer to the NORTH_STAR architecture
by separating concerns within the terminal emulator logic.
